Get rid of various unnecessary warnings in logs (#2981)
* Watch explicit event for channel actor termination

We use an explicit event in the `Peer` when watching for child channel
actors being terminated. This should get rid of the warnings in the logs
about the `Terminated` event not being handled, or at least provide us
more hints of which actors aren't properly being tracked.

* Fix disconnect response in channel actors

We weren't watching for the right event, which creates a lot of log
lines saying that the `Disconnecting` event is unhandled.

* Ignore closing negotiation if closing or closed

If we receive an outdated closing message, we ignore it without creating
a warning in the logs.

* Ignore HTLC settlement commands while disconnected

We ignore HTLC settlement commands while we're disconnected, they will
be retried once the channel has been reestablished. We also ignore
commands asking us to sign the latest state or update fees.