Verify maven dependency checksums
Maven build process will compute checksums for our dependencies and verify them against a local "trusted checksums" file, and fail the build
if the do not match. This process is local and independent of checksums stored and downloaded from maven repositories.

SNAPSHOT dependencies are not verified (but released versions should never depend on SNAPSHOT dependencies).