Clarify recipient blinded path fee deduction
We clarify that the previously named "hiddenFee" is actually the blinded
path fee that is being paid by the recipient (by deducing it from the
amount it plans on receiving). We thus rename variables and add docs.

We also revert the validation flow: instead of generating a payment part
for that fee based on the difference between the paid amount and the
onion value, we generate that payment part based on the blinded path fee
we recorded in the `path_id` and validate that it matches what the payer
is sending.

We revert some of the changes to the `MultiPartHandler` which aren't
necessary and do some clean-up of some unused code paths.