Stop using `QuickLens` in non-test code
QuickLens is handy when modifying deeply nested structures, but since it
internally uses reflection, it may break deterministic builds. We thus
remove all usage of this library in non-test code. We didn't really need
it anyway, it's mostly in tests that we do deep modifications of channel
internal state to be able to test edge cases.