Improve Bolt12 offer APIs (#3037)
The `listoffers` RPC previously exposed raw TLV data and omitted the
`createdAt` and `disabledAt` information, which made it quite useless.
We now return all of the useful offer data, which makes it easier for
node operators to manage their offers.

If we register the same offer twice, we previously crashed because
we were not handling primary key constraints failures. We now
correctly handle them and report an error.