Fix flaky GossipIntegrationSpec test (#3039)
If Alice receives `ExternalChannelSpentTriggered` after validating the local channel update for the splice, then the original funding tx will not be in the `spentChannels` list and the splice will be added to the routing graph as a new channel.

When `ExternalChannelSpentTriggered` is received, it will add Alice's funding tx to `spentChannels` and cause the assert in issue 3033.

This is unlikely to occur in the wild and if it does, the original funding tx will be removed when the splice tx that spends it confirms.