Support p2tr bitcoin wallet (#3026)
* Add an address type parameter to our bitcoin core RPC client

We defined an new AddressType type that can optionally be passed to calls tp `getReceiveAddress()`.
This type can represent p2wpkh or p2tr addresses.

* Add a "chain hash" property to the bitcoin rpc client

* Replace getP2wpkhPubkeyHashForChange() with getChangePublicKeyScript()

We should not hardcode the type of change output that we want but instead use the default change type configured in bitcoin core.

* Simplify SingleKeyOnChainWallet

Use a key manager to generate a local address and sign transactions (instead of signing them manually).

* LocalOnChainKeyManager: support signing BIP86 transactions

* make InteractiveTxBuilder compatible with p2tr wallets

If our wallets adds p2tr inputs, to sign them we need to know all the utxos spent by the tx we're building, not just the one spent by the tx we're signing.

* Cache both onchain public key and public key scripts

* Test bitcoin core change output type selection

If the -changetype option is not set, then bitcoin core will fund transactions with inputs that match transaction's outputs (i.e it will add p2wpkh change outputs if the tx sends to p2wpkh outputs, p2tr change outputs if it sends to p2tr outputs...).

* Integration tests: change default bitcoin core address type to p2tr

* Rework `OnChainAddressRefresher`

* Reformat / clean-up code and comments

 * Improve tests

* sendonchain: unlock utxos if publishing fails

* Add signing test with mixed p2wpkh/p2tr inputs and an eclair-backed bitcoin core wallet

---------

Co-authored-by: Bastien Teinturier <31281497+t-bast@users.noreply.github.com>