Optional `payment_secret` in trampoline outer payload
If a payer or intermediate trampoline node doesn't need MPP to reach
the next trampoline node, they may omit the `payment_secret` field
from the outer onion payload and send a single HTLC.

We made `payment_secret` mandatory for final packets, since we set
`payment_secret` as required in our Bolt 11 invoices. This lets us
use the same code when receiving a payment and when receiving a
trampoline relay request. Making the `payment_secret` field optional
in `PaymentOnion.FinalPayload.Standard` has too much of an impact on
the codebase: I instead chose a somewhat hacky version where we simply
insert a dummy `payment_secret` in the trampoline case when it is not
provided.