Relay non-blinded failure from wallet nodes
When using Bolt12 with a wallet node that is directly connected to us,
we don't always need to return an unreadable failure upstream. If the
wallet node sent us a normal failure, we relay it upstream instead of
overriding it: they chose to reveal that failure to the payer so we
should respect that.

We make sure we don't leak a `channel_update` (since the channel was
unannounced), and the wallet should encrypt the failure with their
blinded node_id, so this shouldn't reveal information that wasn't
already available in the invoice.