Use package relay for anchor force-close (#2963)
We use bitcoin core's `submitpackage` RPC to publish commit txs with
their anchor transaction. We do this for both local and remote commit
txs, and prefer the remote commit when it is available.

Thanks to package relay, this allows the commit and anchor package to
propagate even if the commit feerate is below the mempool minimum
feerate, as long as enough nodes run bitcoin core v28 or higher.