ChannelKeyManager: add optional list of spent outputs to sign() methods (#3047)
* ChannelKeyManager: add a map of spent outputs to sign() methods

This is needed to sign inputs that spend taproot outputs. These outputs are typically wallet inputs added to a transaction to bump it.

* Stricter requirements on input details for signing (#3052)

We verify that details about all inputs are provided to the `sign` function. While this isn't mandatory for segwit v0, it ensures that all of our existing tests exercise this codepath and reduces the risk that we forget to provide some wallet inputs, which would result in an invalid signature which would be hard to investigate.

With this change, some of the unit tests started failing, which showed that we weren't correctly setting wallet inputs in the fee-bumping case in `ReplaceableTxFunder`, which we've fixed.

We also add a test in `TransactionsSpec.scala` to verify that signing fails when details about some inputs are missing.

---------

Co-authored-by: Bastien Teinturier <31281497+t-bast@users.noreply.github.com>