Refactor output redeem type from `TransactionWithInputInfo`
We refactor `TransactionWithInputInfo` to extract the redeem information
from the `InputInfo` class. This is necessary for taproot channels (and
other future segwit versions) to easily integrate into our transaction
class hierarchy: this way we can explicitly specify which kind of redeem
paths are possible for each transaction type. For example, we can now
express that a `CommitTx` uses either a segwit v0 input or a taproot
input that we can spend via the key path.

Note that when we'll later introduce v3 commitments, we may have txs
that can be spent:

- via segwit v0 for non-taproot channels
- via a taproot script path for taproot channels
- AND via a taproot key path for v3 taproot channels

This is why we cannot simply pick a single one of the taproot redeem
paths, we may need to support multiple of them in the future.

The core of this PR is just the following small type changes, that
should be reviewed first:

- we introduce the `RedeemInfo` trait
- we remove the redeem information from `InputInfo`, which now becomes
  a trivial container for `OutPoint` + `TxOut`
- we update `CommitmentOutputLink` to use an abstract `RedeemInfo` in
  the `CommitmentOutput`, which will make it easy to add taproot
  outputs without modifying the `CommitmentOutputLink` type
- we take the `commitmentFormat` parameter when creating transactions,
  which makes it easy to pattern match on it to produce the right redeem
  information

Note that this highlighted that we unnecessarily recomputed HTLC scripts
when creating claim-htlc transactions: we can directly leverage the
existing `CommitmentOutput`s. I wondered if we could do this for every
transaction creation, but we can't because we need to handle spending
revoked commitments, for which we don't have a `CommitmentSpec` object
and thus cannot create the set of `CommitmentOutput`s.