Clean-up `TransportHandler`
The `TransportHandler` is a very old actor that we never refactored much
and needed a bit of clean-up. There is no reason to make it generic, it
only supports sending lightning messages on the wire. If we ever need to
make it generic in the future, we can easily do it, but for simplicity
it should only handle lightning messages for now.