Introduce `CommitmentKeys` class
We introduce dedicated classes to hold the keys used for either a local
commit transaction or a remote commit transaction. These keys directly
map to the keys described in Bolt 3. The goal is to generate those keys
only once instead of what we're currently doing, where we duplicate a
lot of code to partially generate keys in various places, which is very
error-prone: incorrectly using the wrong public or private key can lead
to loss of funds! It also removes the burden of inverting local/remote
in many places when we're spending a remote commitment: this only has
to happen once, in the `publicKeys` field of `RemoteCommitmentKeys`.

In this first commit, we only use this to generate commit txs, htlc txs
and claim-htlc txs. We generate these classes next to where the keys
were previously derived, to make it easier to review the diff. In the
next commit, we will create helper functions to generate those keys
based on the `Commitment` object.