Create all transactions from commitment keys
We now create all commitment-related transactions using commitment keys
instead of explicitly using a subset of the public keys used.

Note how this highlights that many tests in `TransactionsSpec.scala`
were actually using incorrect keys, and thus not really testing the
correct behavior. This is a very good proof that making this simpler
is a good choice, because that kind of issues can also happen in non
test code (and has happened in the past already).