Remove signing functions from `ChannelKeyManager`
We were previously going through the `ChannelKeyManager` to sign txs
just to end up calling `TransactionWithInputInfo.sign` after applying
some key derivation. This was confusing and is now unnecessary: we can
directly call `TransactionWithInputInfo.sign` since the new commitment
keys directly contain the correctly tweaked keys.

The `ChannelKeyManager`'s responsibility is now simply to derive the
keys and ensure that different channels use different derivation paths.
Unfortunately, we still need to call it in a few different places to
obtain the channel key path and then derive some of our keys. This can
be improvied: we will go further in the following commits to simplify
this.