Split `ChannelKeyManager` and `ChannelKeys`
The `ChannelKeyManager` previously had two distinct roles:

- generating random funding key paths for new channels
- generating channel keys from a funding key path

This was confusing and forced all the code that wanted to derive channel
keys to always first get the channel key path from the key manager and
then call the function it was interested in with this channel key path
as argument.

We remove this confusion by having the `ChannelKeyManager` only generate
funding key paths for new channels and initialize `ChannelKeys` based on
a given funding key path. The new `ChannelKeys` object provides direct
access to derivation functions for all the keys necessary. Each channel
has its own `ChannelKeys` object that only lets it derive keys for its
own usage.

Reviewers should start by reviewing the changes to `ChannelKeyManager`
and `LocalChannelKeyManager`: the rest is mostly mechanical changes to
use this new architecture.

We also refactor a bit the functions in `Commitments.scala` to derive
commitment keys as early as possible and then provide them as arguments
to the child functions.