Move `Generators` functions inside `ChannelKeys`
We move the commitment key derivation logic inside `ChannelKeys`, which
makes it easier to understand the various types of derivation involved
in channel keys.

We also explicitly pattern match on `commitmentFormat` when deciding
whether the payment key needs to be tweaked for each commitment or not,
which makes it more future-proof. It is a bit confusing because we did
not introduce a `commitmentFormat` for `option_static_remotekey`, which
applies either to anchor outputs or the default commitment format when
the corresponding channel feature is turned on.