Introduce solvers to simplify how we spend commit tx transactions
We update CommitmentOutput to include the actual output that it represents in the commit tx, and the redeem script to spend it.
We also include 2nd level HTLC transactions when they are not trimmed.
We remove CommitmentOutputLink which is now useless, and simplify how we create commitment and HTLC transactions.
We add solvers to create and sign spending transactions, and remove makeXXXTx() aand addSigs() methods.