Refactor commitment outputs
We update CommitmentOutput to include the actual output that it represents in the commit tx, and the redeem script to spend it.
We also include 2nd level HTLC transactions when they are not trimmed.
We remove CommitmentOutputLink which is now useless, and simplify how we create commitment and HTLC transactions.