Move `addSigs` into each dedicated class
Instead of having multiple overloads for `addSigs`, we move each one of
them inside the corresponding `TransactionWithInputInfo`. This is part
of an effort to group the logic belonging to each type of transaction
in the same place to make it easier to review and use.

This is a purely mechanical refactoring, where code has simply been cut
and pasted. The only functional change is for `HtlcPenaltyTx`, where we
incorrectly used the witness for a `MainPenaltyTx`, which was actually
smaller. We were thus under-estimating the fees we had to pay to reach
the desired feerate, which isn't much of an issue, but it's better to
use the correct witness for consistency.