Simplify channel keys management (#3064)
Our previous architecture for channel key management was designed when
we wanted the ability to use an external signer that would protect the
private keys used by `eclair`: this meant we couldn't directly deal
with private keys and forced us to have a more complex, indirect
architecture which made it harder to understand how key derivation is
done for commitment keys and forced us to partially derive some of the
keys in many different places, which was very error-prone (when should
we provide a `payment_key` instead of a `delayed_payment_key`?).

Since then, we've spent years working on lightning-specific HSMs before
VLS was even a thing, and ended up concluding that this was a dead end
and it made much more sense to put the whole lightning node inside a
secure enclave: https://acinq.co/blog/securing-a-100M-lightning-node

But we haven't refactored our key management code since then. Now that
we're working on Taproot, which adds more complexity to transactions,
it's a good time to re-work that part of the code and make it simpler.
This refactoring took a while and required a few iterations to get
right. I think the result is much better than what we previously had:

- all of the key derivation is contained in `ChannelKeys`, which makes
  it easier to grasp
- each channel has its own `ChannelKeys` instance, containing only what
  it needs
- from a `ChannelKeys` instance and a per-commitment point, we can
  derive a set of `CommitmentKeys` that allow creating local/remote
  commitment and HTLC transactions
- our force-close logic (in `Helpers.scala`) does *not* need to derive
  keys manually and can instead rely on a `CommitmentKeys` instance
- script and witness creation also take a `CommitmentKeys` instance
  instead of specific public keys, which was extremely error-prone
  because we relied on naming instead of types to ensure that the
  right public key was provided

Importantly, this is only a first step, that will allow us to also
greatly simplify our whole `TransactionWithInputInfo` architecture,
which is also starting to show its age.