Remove `LegacyClaimHtlcSuccess` transaction class
We have this legacy class in the `ClaimHtlcTx` hierarchy for legacy
cases that didn't store the `paymentHash` with the transaction. This
was introduced to ensure that nodes who upgraded while a force-close
was happening would still be able to finish force-closing the channel
and get their funds back.

This was introduced in #2101 which was released in eclair v0.7.0 on
01/02/2022. Since we've made several important bug fixes since then,
it should be safe to assume that we don't care about node operators
who update from this very old version while a channel is closing, the
remote commit got confirmed, and some HTLCs must be claimed with the
preimage on-chain.