Add missing data to `TransactionWithInputInfo` classes
Most of our instances of `TransactionWithInputInfo` only contained the
input (`InputInfo` instance) and the transaction. This worked because
we stored the (segwit v0) redeem script in the `InputInfo` instance.
But we can actually get rid of that if we store with each transaction
the parameters that allow re-creating the scripts at runtime.

We don't need many parameters, so this is quite simple to add:

- delayed txs simply need the `to_self_delay` value used
- HTLC txs simply need the `payment_hash` and `htlc_expiry`

Based on that we can trivially re-build scripts and redeem info at
runtime using `ChannelKeys`. This ensures that we will only need the
`commitmentFormat` to know the exact script that is being used, instead
of having to encode different redeem information in `InputInfo`. This
will also reduce the size of our channel data, because storing redeem
info can be removed.

In order to handle data that was encoded before this change, we:

- update `ChannelCodecs4.scala`: this is quite a simple change, but this
  file is getting messy since we made many changes to transaction codecs
  so please review carefully
- when re-connecting a channel, if it contains transactions that are
  missing some data, we simply re-create those transactions based on
  the commitment data