Refactor transaction signing
In this commit, we somewhat heavily re-design `Transactions.scala`. The
main change is that we stop exposing the default `sign` function that
was too abstract for callers, and have each `TransactionWithInputInfo`
expose its dedicated `sign` function, with simpler arguments depending
on each transaction type.

One very important thing to note is that there is a heavy distinction
between transactions that require both a local and a remote signature
(commit-tx, splice-tx, closing-tx and htlc-tx) and transactions that
only need a local signature. This means we can completely remove the
`addSigs` step for transactions that only need a local signature, and
directly create signed signatures, which simplifies the calling code.

This also lets us better encapsulate the `extraUtxos` parameter, which
only makes sense for txs for which we add wallet inputs, which is only
a small subset of our transactions.

The previous commits ensured that we store the data we need for each
transaction type to allow re-deriving and signing everything based
only on the following runtime data:

- the commitment format
- the commitment funding keys
- the commitment keys

This makes the calling code much less error-prone and removes the need
to store the redeem script in the `InputInfo` class, which wouldn't
have worked for taproot transactions. Note that for simplicity's sake,
we currently keep the field (to avoid a painful codecs migration) since
we will introduce v5 codecs when implementing taproot channels, which
will let us clean up this inconsistency.

We also move each transaction's creation function in its companion
object: this makes the diff a bit harder to review, but ensures that
each transaction's implementation is well encapsulated, which makes
it easy to review that each transaction type is correctly implemented.

Instead of using placeholder signatures when evaluating the weight of
transactions, we actually sign them. This means we sign twice, but is
not a performance issue in this case because it only happens once when
we initiate a force-close, and simplifies the code.

We also introduce a `RedeemInfo` trait which makes the code forward
compatible with taproot channels. The changes to `Transactions.scala`
are a lot to review, but unfortunately I didn't find a good way to
split them while having something consistent and all the unit tests
passing...