Dedicated trait for `ChannelSpendTransaction`s
Transactions that spend the channel funding output don't behave the
same way as transactions that can be unilaterally signed, so it makes
sense to use different traits. This lets us better scope our shared
code and remove unused parameters (e.g. sighash is much simpler for
transactions spending the channel funding output and the concept of
`TxOwner` is unnecessary).

We also take this opportunity to explicitly separate the two distinct
signing cases for `ChannelSpendTransaction`: 2-of-2 multisig and musig2,
which create different signature types. By pattern matching on the
commitment format, callers should decide whether they want to produce
an individual signature or a partial signatures. This may need some
rework to fully adapt to taproot, but it has the benefit of exposing
which part of the code should be concerned about the type of sigs it
creates.