Publish 3rd-stage penalty txs on tx confirmation
In case of a revoked close, we previously published 3rd-stage penalty
txs as soon as we saw our peer's HTLC transactions in our mempool.
This isn't necessary and may actually help them get confirmed since
it is essentially a CPFP. We now instead wait for the revoked HTLC
transaction to be confirmed before publishing a penalty transaction
claiming its outputs.