Refactor some closing helper functions (#3089)
* Move `confirmationTarget` to closing helper function

We move functions that compute the confirmation target outside of
`LocalCommitPublished` and `RemoteCommitPublished` to helper functions
in the `Closing` object.

We also slightly refactor `doPublish` functions to remove the pattern
matching on the commitment format, which was actually unnecessary since
we introduced the `redeemableHtlcTxs` function.

* Refactor anchor tx creation

This is a trivial refactoring where we explicitly create a single anchor
transaction instead of replacing an empty list. We also clean up
comments and variable names in force-close helpers.

* Refactor upstream trimmed or timed out HTLCs

We stop relying on `LocalCommitPublished` and `RemoteCommitPublished` to
identify when HTLCs should be failed upstream because they are either
trimmed or their timeout transaction is confirmed. We instead rely on
commitment data for local commitments, and re-create the commit outputs
on-the-fly for remote commitments.

* Publish 3rd-stage penalty txs on tx confirmation

In case of a revoked close, we previously published 3rd-stage penalty
txs as soon as we saw our peer's HTLC transactions in our mempool.
This isn't necessary and may actually help them get confirmed since
it is essentially a CPFP. We now instead wait for the revoked HTLC
transaction to be confirmed before publishing a penalty transaction
claiming its outputs.